Resolved conflicts: kept dev's version for all test files (already merged via PR #160).

+++ LivingRoots/Domain/Models/CompostingBinStateModel.cs
@@ -18,8 +18,8 @@ public class CompostingBinStateModel
     /// <summary>Qualified item ID of the input waste item (null if empty).</summary>
     public string? InputItemId { get; set; }
 
-    /// <summary>Game time in minutes when waste was added (null if empty).</summary>
-    public long? InputTimestamp { get; set; }
+    /// <summary>Game day when waste was added (null if empty).</summary>
+    public int? InputTimestamp { get; set; }
 
     /// <summary>Current maturation level (1-5), determines output multiplier.</summary>
     public int MaturationLevel { get; set; } = 1;
